Renamed server folder to be src
