+++ src/bin/server.js
